refactor: introduce `dns_mapping` module
